refactor: 채팅방 훅 분리 + DB 테이블 type 분리 #10
