+++ src/types/chatroom.ts
@@ -1,0 +1,6 @@
+import { GenericTables } from "@/types/supabase.types";
+
+/**
+ * public.chatroom 테이블 row
+ */
+export type Room = GenericTables<"chatroom">;
