+++ src/types/message.ts
@@ -1,0 +1,6 @@
+import { GenericTables } from "@/types/supabase.types";
+
+/**
+ * public.message 테이블 row
+ */
+export type Message = GenericTables<"message">;
